requirements.txt
# FQ51BBS Python Dependencies
# Install with: pip install -r requirements.txt

# Core - Meshtastic communication
meshtastic>=2.3.0
pubsub>=4.0.3

# Cryptography
argon2-cffi>=23.1.0          # Password hashing (Argon2id)
cryptography>=41.0.0          # ChaCha20-Poly1305 encryption

# Configuration
tomli>=2.0.0                  # TOML parsing (Python <3.11)
toml>=0.10.2                  # TOML writing

# Web reader (optional)
flask>=3.0.0

# Development/Testing (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
